Updated to make error messages on data set installer more clear

+++ SanteDB.Persistence.Data/Services/AdoDatasetInstallerService.cs
@@ -32,6 +32,7 @@ using SanteDB.Persistence.Data.Configuration;
 using SanteDB.Persistence.Data.Model.Sys;
 using System;
 using System.Collections.Generic;
+using System.Data.Common;
 using System.Linq;
 using System.Reflection;
 using System.Security.Cryptography;
@@ -234,9 +235,17 @@ namespace SanteDB.Persistence.Data.Services
 
                                         break;
                                     case DataDelete dd:
-                                        persistenceService.Delete(context, dd.Element.Key.Value, DataPersistenceControlContext.Current?.DeleteMode ?? this.m_configuration.DeleteStrategy);
+                                        if (persistenceService.Exists(context, dd.Element.Key.Value))
+                                        {
+                                            persistenceService.Delete(context, dd.Element.Key.Value, DataPersistenceControlContext.Current?.DeleteMode ?? this.m_configuration.DeleteStrategy);
+                                        }
                                         break;
                                 }
+                            }
+                            catch(DbException e)
+                            {
+                                this.m_tracer.TraceError("Installing {0} (#{1} in dataset) failed", itm, i);
+                                throw e.TranslateDbException();
                             }
                             catch (Exception e)
                             {
@@ -247,6 +256,7 @@ namespace SanteDB.Persistence.Data.Services
                                 }
                                 else
                                 {
+                                    this.m_tracer.TraceError("Error applying dataset item {0} - {1}", i, e.ToHumanReadableString());
                                     throw;
                                 }
                             }
@@ -294,7 +304,7 @@ namespace SanteDB.Persistence.Data.Services
                 }
                 catch (Exception e)
                 {
-                    throw new DataPersistenceException(String.Format(ErrorMessages.BUNDLE_PERSISTENCE_ERROR, dataset.Id, "X"), e);
+                    throw new DataPersistenceException(String.Format(ErrorMessages.BUNDLE_PERSISTENCE_ERROR, "X", dataset.Id), e);
                 }
             }
         }
